Добавит абстрактный класс и наследование
