+++ src/view/films-list-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createFilmsListTemplate = () => (
   `<section class="films-list">
@@ -9,22 +9,8 @@ const createFilmsListTemplate = () => (
   </section>`
 );
 
-export default class FilmsListView {
-  #element = null;
-
+export default class FilmsListView extends AbstractView {
   get template() {
     return createFilmsListTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
