+++ src/view/films-view.js
@@ -1,25 +1,11 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createFilmsTemplate = () => (
   '<section class="films"></section>'
 );
 
-export default class FilmsView {
-  #element = null;
-
+export default class FilmsView extends AbstractView {
   get template() {
     return createFilmsTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
