+++ src/view/more-button-view.js
@@ -1,25 +1,11 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createMoreButtonTemplate = () => (
   '<button class="films-list__show-more">Show more</button>'
 );
 
-export default class MoreButtonView {
-  #element = null;
-
+export default class MoreButtonView extends AbstractView {
   get template() {
     return createMoreButtonTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
